+++ lib/ui/widgets/MapWidget.dart
@@ -94,10 +94,6 @@ class MapController {
     return _channel.invokeMethod('enableMyLocation', enable);
   }
 
-  Future<void>enableLevels(bool enable) async {
-    return _channel.invokeMethod('enableLevels', enable);
-  }
-
   Future<void> viewPoi(Map<String, dynamic>? target) async {
     return _channel.invokeMethod('viewPoi', {'target': target});
   }